de-abstraction of components, fixed header uniformity
