+++ src/components/eventsPage/eventCard.tsx
