+++ src/components/eventsPage/eventDrawer.tsx
